Turn the launcher into the toolset
