+++ android/app/src/main/AndroidManifest.xml
@@ -37,9 +37,11 @@
         android:supportsRtl="false">
 
         <activity
-            android:name="LauncherActivity"
-            android:alwaysRetainTaskState="true"
+            android:name=".ToolsetActivity"
             android:exported="true"
+            android:label="@string/activity_toolset_label"
+            android:launchMode="singleInstance"
+            android:taskAffinity="org.fheroes2.ToolsetActivityTask"
             android:theme="@style/Theme.Material3.DayNight">
 
             <intent-filter>
@@ -49,13 +51,21 @@
         </activity>
 
         <activity
-            android:name="GameActivity"
+            android:name=".GameActivity"
+            android:alwaysRetainTaskState="true"
             android:configChanges="layoutDirection|locale|orientation|uiMode|screenLayout|screenSize|smallestScreenSize|keyboard|keyboardHidden|navigation"
-            android:exported="false"
+            android:exported="true"
+            android:launchMode="singleInstance"
             android:preferMinimalPostProcessing="true"
-            android:process=":GameActivity"
+            android:process=":GameActivityProcess"
             android:screenOrientation="sensorLandscape"
+            android:taskAffinity="org.fheroes2.GameActivityTask"
             android:theme="@android:style/Theme.NoTitleBar.Fullscreen">
+
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
 
             <!-- Let Android know that we can handle some USB devices and should receive this event -->
             <intent-filter>
