+++ android/app/src/main/res/layout/activity_toolset.xml
@@ -6,48 +6,34 @@
     android:layout_height="match_parent">
 
     <ImageView
-        android:id="@+id/activity_launcher_game_logo_img"
+        android:id="@+id/activity_toolset_game_logo_img"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginBottom="32dp"
-        android:contentDescription="@string/activity_launcher_game_logo_img_content_description"
+        android:contentDescription="@string/activity_toolset_game_logo_img_content_description"
         android:src="@mipmap/ic_launcher"
-        app:layout_constraintBottom_toTopOf="@+id/activity_launcher_inactivity_timer_pb"
+        app:layout_constraintBottom_toTopOf="@+id/activity_toolset_game_status_lbl"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent" />
 
-    <ProgressBar
-        android:id="@+id/activity_launcher_inactivity_timer_pb"
-        style="@android:style/Widget.ProgressBar.Horizontal"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="16dp"
-        android:minWidth="256dp"
-        android:visibility="gone"
-        app:layout_constraintBottom_toTopOf="@+id/activity_launcher_game_status_lbl"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_chainStyle="packed" />
-
     <TextView
-        android:id="@+id/activity_launcher_game_status_lbl"
+        android:id="@+id/activity_toolset_game_status_lbl"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginBottom="16dp"
         android:gravity="center"
         android:maxWidth="512dp"
         android:padding="8dp"
-        android:text="@string/activity_launcher_game_status_lbl_text"
+        android:text="@string/activity_toolset_game_status_lbl_text"
         android:visibility="gone"
-        app:layout_constraintBottom_toTopOf="@+id/activity_launcher_run_game_btn"
+        app:layout_constraintBottom_toTopOf="@+id/activity_toolset_run_game_btn"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/activity_launcher_inactivity_timer_pb"
+        app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_chainStyle="packed" />
 
     <Button
-        android:id="@+id/activity_launcher_run_game_btn"
+        android:id="@+id/activity_toolset_run_game_btn"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginBottom="8dp"
@@ -55,31 +41,31 @@
         android:minWidth="256dp"
         android:onClick="runGameButtonClicked"
         android:padding="8dp"
-        android:text="@string/activity_launcher_run_game_btn_name"
-        app:layout_constraintBottom_toTopOf="@+id/activity_launcher_extract_homm2_resources_btn"
+        android:text="@string/activity_toolset_run_game_btn_name"
+        app:layout_constraintBottom_toTopOf="@+id/activity_toolset_extract_homm2_assets_btn"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/activity_launcher_game_status_lbl"
+        app:layout_constraintTop_toBottomOf="@+id/activity_toolset_game_status_lbl"
         app:layout_constraintVertical_chainStyle="packed" />
 
     <Button
-        android:id="@+id/activity_launcher_extract_homm2_resources_btn"
+        android:id="@+id/activity_toolset_extract_homm2_assets_btn"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginBottom="8dp"
         android:enabled="false"
         android:minWidth="256dp"
-        android:onClick="extractHoMM2ResourcesButtonClicked"
+        android:onClick="extractHoMM2AssetsButtonClicked"
         android:padding="8dp"
-        android:text="@string/activity_launcher_extract_homm2_resources_btn_name"
-        app:layout_constraintBottom_toTopOf="@+id/activity_launcher_download_homm2_demo_btn"
+        android:text="@string/activity_toolset_extract_homm2_assets_btn_name"
+        app:layout_constraintBottom_toTopOf="@+id/activity_toolset_download_homm2_demo_btn"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/activity_launcher_run_game_btn"
+        app:layout_constraintTop_toBottomOf="@+id/activity_toolset_run_game_btn"
         app:layout_constraintVertical_chainStyle="packed" />
 
     <Button
-        android:id="@+id/activity_launcher_download_homm2_demo_btn"
+        android:id="@+id/activity_toolset_download_homm2_demo_btn"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginBottom="16dp"
@@ -87,27 +73,27 @@
         android:minWidth="256dp"
         android:onClick="downloadHoMM2DemoButtonClicked"
         android:padding="8dp"
-        android:text="@string/activity_launcher_download_homm2_demo_btn_name"
-        app:layout_constraintBottom_toTopOf="@+id/activity_launcher_background_task_pb"
+        android:text="@string/activity_toolset_download_homm2_demo_btn_name"
+        app:layout_constraintBottom_toTopOf="@+id/activity_toolset_background_task_pb"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/activity_launcher_extract_homm2_resources_btn"
+        app:layout_constraintTop_toBottomOf="@+id/activity_toolset_extract_homm2_assets_btn"
         app:layout_constraintVertical_chainStyle="packed" />
 
     <ProgressBar
-        android:id="@+id/activity_launcher_background_task_pb"
+        android:id="@+id/activity_toolset_background_task_pb"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginBottom="8dp"
         android:visibility="gone"
-        app:layout_constraintBottom_toTopOf="@+id/activity_launcher_last_task_status_lbl"
+        app:layout_constraintBottom_toTopOf="@+id/activity_toolset_last_task_status_lbl"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/activity_launcher_download_homm2_demo_btn"
+        app:layout_constraintTop_toBottomOf="@+id/activity_toolset_download_homm2_demo_btn"
         app:layout_constraintVertical_chainStyle="packed" />
 
     <TextView
-        android:id="@+id/activity_launcher_last_task_status_lbl"
+        android:id="@+id/activity_toolset_last_task_status_lbl"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:freezesText="true"
@@ -119,6 +105,6 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/activity_launcher_background_task_pb"
+        app:layout_constraintTop_toBottomOf="@+id/activity_toolset_background_task_pb"
         app:layout_constraintVertical_chainStyle="packed" />
 </androidx.constraintlayout.widget.ConstraintLayout>
